Fix sql type

+++ cluster/resources/schemas/dbscripts/postgresql/cluster-15.24-15.25.sql
@@ -1,4 +1,4 @@
 ALTER TABLE cluster.clusterJobs ADD clusterAccount varchar(1000);
-ALTER TABLE cluster.clusterJobs ADD duration double;
+ALTER TABLE cluster.clusterJobs ADD duration double precision;
 ALTER TABLE cluster.clusterJobs ADD cpuUsed int;
 ALTER TABLE cluster.clusterJobs ADD gpuUsed int;
